src/functions/assemblyai-webhook-cloud-func/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via httpx
assemblyai==0.37.0
    # via -r requirements.in
blinker==1.9.0
    # via flask
cachetools==5.5.1
    # via google-auth
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via flask
flask==3.1.0
    # via -r requirements.in
google-api-core==2.24.1
    # via
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-storage
google-api-python-client==2.161.0
    # via -r requirements.in
google-auth==2.38.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-cloud-core==2.4.1
    # via
    #   -r requirements.in
    #   google-cloud-storage
google-cloud-storage==3.0.0
    # via -r requirements.in
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.67.0
    # via google-api-core
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via assemblyai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.5
    # via flask
lxml==5.3.1
    # via python-docx
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
proto-plus==1.26.0
    # via google-api-core
protobuf==5.29.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.10.6
    # via assemblyai
pydantic-core==2.27.2
    # via pydantic
pyparsing==3.2.1
    # via httplib2
python-docx==1.1.2
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-storage
rsa==4.9
    # via google-auth
sniffio==1.3.1
    # via anyio
typing-extensions==4.12.2
    # via
    #   anyio
    #   assemblyai
    #   pydantic
    #   pydantic-core
    #   python-docx
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.3.0
    # via requests
websockets==15.0
    # via assemblyai
werkzeug==3.1.3
    # via flask
